Ajout de Dwell-Time Fixation Zero Duration au jeu "Bubble"
