+++ gazeplay-games/src/main/java/net/gazeplay/games/bubbles/Bubble.java
@@ -5,6 +5,7 @@ import javafx.event.Event;
 import javafx.event.EventHandler;
 import javafx.geometry.Dimension2D;
 import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.image.Image;
 import javafx.scene.input.MouseEvent;
 import javafx.scene.paint.Color;
@@ -22,6 +23,8 @@ import net.gazeplay.commons.utils.games.ImageLibrary;
 import net.gazeplay.commons.utils.games.ImageUtils;
 import net.gazeplay.commons.utils.games.Utils;
 import net.gazeplay.commons.utils.stats.Stats;
+import net.gazeplay.components.Portrait;
+import net.gazeplay.components.RandomPositionGenerator;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,17 +33,12 @@ import java.util.List;
  * Created by schwab on 28/08/2016.
  */
 @Slf4j
-public class Bubble extends Parent implements GameLifeCycle {
+public class Bubble implements GameLifeCycle {
 
     public static final String DIRECTION_TOP = "toTop";
     public static final String DIRECTION_BOTTOM = "toBottom";
     public static final String DIRECTION_LEFT = "toLeft";
     public static final String DIRECTION_RIGHT = "toRight";
-
-    private static final int maxTimeLength = 7;
-    private static final int minTimeLength = 4;
-
-    private static final int nbFragments = 10; // number of little circles after explosion
 
     private final IGameContext gameContext;
 
@@ -48,42 +46,28 @@ public class Bubble extends Parent implements GameLifeCycle {
 
     private final Stats stats;
 
-    private final ImageLibrary imageLibrary;
+    private Target portrait;
 
-    private List<Circle> fragments;
-
-    private EventHandler<Event> enterEvent;
-
-    private final BubblesGameVariant direction;
+    private final BubblesGameVariant gameVariant;
 
     private final ReplayablePseudoRandom randomGenerator;
-    private final ReplayablePseudoRandom randomFragmentsGenerator = new ReplayablePseudoRandom();;
 
-    public Bubble(final IGameContext gameContext, final BubbleType type, final Stats stats, final BubblesGameVariant direction) {
+    public Bubble(final IGameContext gameContext, final BubbleType type, final Stats stats, final BubblesGameVariant gameVariant) {
+        super();
         this.gameContext = gameContext;
         this.type = type;
         this.stats = stats;
-        this.direction = direction;
-        gameContext.startTimeLimiter();
-        gameContext.startScoreLimiter();
-
+        this.gameVariant = gameVariant;
         this.randomGenerator = new ReplayablePseudoRandom();
         this.stats.setGameSeed(randomGenerator.getSeed());
-
-        imageLibrary = ImageUtils.createImageLibrary(Utils.getImagesSubdirectory("portraits"), randomGenerator);
     }
 
-    public Bubble(final IGameContext gameContext, final BubbleType type, final Stats stats, final BubblesGameVariant direction, double gameSeed) {
+    public Bubble(final IGameContext gameContext, final BubbleType type, final Stats stats, final BubblesGameVariant gameVariant, double gameSeed) {
         this.gameContext = gameContext;
         this.type = type;
         this.stats = stats;
-        this.direction = direction;
-        gameContext.startTimeLimiter();
-        gameContext.startScoreLimiter();
-
+        this.gameVariant = gameVariant;
         this.randomGenerator = new ReplayablePseudoRandom(gameSeed);
-
-        imageLibrary = ImageUtils.createImageLibrary(Utils.getImagesSubdirectory("portraits"), randomGenerator);
     }
 
     void initBackground(boolean useBackgroundImage) {
@@ -111,226 +95,23 @@ public class Bubble extends Parent implements GameLifeCycle {
 
     @Override
     public void launch() {
-        this.getChildren().clear();
         initBackground(true);
-        gameContext.getChildren().add(this);
-        gameContext.setLimiterAvailable();
+        final RandomPositionGenerator randomPositionGenerator = gameContext.getRandomPositionGenerator();
+        randomPositionGenerator.setRandomGenerator(randomGenerator);
 
-        this.fragments = buildFragments(type);
-
-        this.getChildren().addAll(fragments);
-
-        enterEvent = e -> {
-
-            if (e.getEventType() == MouseEvent.MOUSE_ENTERED || e.getEventType() == GazeEvent.GAZE_ENTERED) {
-
-                enter((Circle) e.getTarget());
-            }
-        };
-
-        gameContext.start();
         for (int i = 0; i < 10; i++) {
-
-            newCircle();
+            portrait = new Target(gameContext, randomPositionGenerator, stats,
+                Portrait.createImageLibrary(randomGenerator), gameVariant, this, randomGenerator, type);
         }
-
-
-        this.fragments = buildFragments(type);
-
-        this.getChildren().addAll(fragments);
-
-        enterEvent = e -> {
-
-            if (e.getEventType() == MouseEvent.MOUSE_ENTERED || e.getEventType() == GazeEvent.GAZE_ENTERED) {
-
-                enter((Circle) e.getTarget());
-            }
-        };
-
-
+        gameContext.setLimiterAvailable();
+        gameContext.getChildren().add(portrait);
         stats.notifyNewRoundReady();
         gameContext.getGazeDeviceManager().addStats(stats);
+        stats.incrementNumberOfGoalsToReach();
     }
 
     @Override
     public void dispose() {
-        enterEvent = e -> {
-        };
-        this.getChildren().clear();
+        gameContext.clear();
     }
-
-    private List<Circle> buildFragments(final BubbleType bubbleType) {
-        final List<Circle> fragments = new ArrayList<>(nbFragments);
-
-        for (int i = 0; i < nbFragments; i++) {
-
-            final Circle fragment = new Circle();
-            fragment.setOpacity(1);
-            fragment.setRadius(20);
-            fragment.setVisible(true);
-            fragment.setCenterX(-100);
-            fragment.setCenterY(-100);
-
-            if (bubbleType == BubbleType.COLOR) {
-                fragment.setFill(new Color(randomFragmentsGenerator.nextDouble(), randomFragmentsGenerator.nextDouble(), randomFragmentsGenerator.nextDouble(), 1));
-            } else {
-                fragment.setFill(new ImagePattern(imageLibrary.pickRandomImage(), 0, 0, 1, 1, true));
-            }
-
-            fragments.add(fragment);
-        }
-
-        return fragments;
-    }
-
-    public void explose(final double centerX, final double centerY) {
-
-        final Timeline goToCenterTimeline = new Timeline();
-        final Timeline timeline = new Timeline();
-
-        for (int i = 0; i < nbFragments; i++) {
-
-            final Circle fragment = fragments.get(i);
-
-            fragment.setCenterX(centerX);
-            fragment.setCenterY(centerY);
-            fragment.setOpacity(1);
-
-            goToCenterTimeline.getKeyFrames().add(new KeyFrame(new Duration(1),
-                new KeyValue(fragment.centerXProperty(), centerX, Interpolator.LINEAR)));
-            goToCenterTimeline.getKeyFrames().add(new KeyFrame(new Duration(1),
-                new KeyValue(fragment.centerYProperty(), centerY, Interpolator.EASE_OUT)));
-            goToCenterTimeline.getKeyFrames().add(new KeyFrame(new Duration(1), new KeyValue(fragment.opacityProperty(), 1)));
-
-            final Dimension2D sceneDimensions = gameContext.getGamePanelDimensionProvider().getDimension2D();
-            final double endXValue = randomFragmentsGenerator.nextDouble() * sceneDimensions.getWidth();
-            final double endYValue = randomFragmentsGenerator.nextDouble() * sceneDimensions.getHeight();
-
-            timeline.getKeyFrames().add(new KeyFrame(new Duration(1000),
-                new KeyValue(fragment.centerXProperty(), endXValue, Interpolator.LINEAR)));
-            timeline.getKeyFrames().add(new KeyFrame(new Duration(1000),
-                new KeyValue(fragment.centerYProperty(), endYValue, Interpolator.EASE_OUT)));
-            timeline.getKeyFrames().add(new KeyFrame(new Duration(1000), new KeyValue(fragment.opacityProperty(), 0)));
-        }
-
-        final SequentialTransition sequence = new SequentialTransition();
-        sequence.getChildren().addAll(goToCenterTimeline, timeline);
-        sequence.play();
-
-        if (randomFragmentsGenerator.nextDouble() > 0.5) {
-            final String soundResource = "data/bubble/sounds/Large-Bubble-SoundBible.com-1084083477.mp3";
-            gameContext.getSoundManager().add(soundResource);
-        } else {
-            final String soundResource = "data/bubble/sounds/Blop-Mark_DiAngelo-79054334.mp3";
-            gameContext.getSoundManager().add(soundResource);
-        }
-    }
-
-    private void enter(final Circle target) {
-
-        final double centerX = target.getCenterX();
-        final double centerY = target.getCenterY();
-
-        gameContext.getGazeDeviceManager().removeEventFilter(target);
-        this.getChildren().remove(target);
-
-        stats.incrementNumberOfGoalsReached();
-
-        gameContext.updateScore(stats,this);
-
-        explose(centerX, centerY); // instead of C to avoid wrong position of the explosion
-
-        this.newCircle();
-    }
-
-    private void newCircle() {
-        final Circle circle = buildCircle();
-        circle.toBack();
-
-        this.getChildren().add(circle);
-        this.gameContext.resetBordersToFront();
-        stats.incrementNumberOfGoalsToReach();
-
-        gameContext.getGazeDeviceManager().addEventFilter(circle);
-
-        circle.addEventFilter(MouseEvent.ANY, enterEvent);
-        circle.addEventHandler(GazeEvent.ANY, enterEvent);
-
-        moveCircle(circle);
-    }
-
-    private Circle buildCircle() {
-
-        final Circle newCircle = new Circle();
-        final Dimension2D screenDimension = gameContext.getGamePanelDimensionProvider().getDimension2D();
-        double maxRadius = Math.min(screenDimension.getWidth()/12,screenDimension.getHeight()/12);
-        double minRadius =  Math.min(screenDimension.getHeight()/30,screenDimension.getWidth()/30);
-        final double radius = (maxRadius - minRadius) * randomGenerator.nextDouble() + minRadius;
-
-        newCircle.setRadius(radius);
-
-        if (type == BubbleType.COLOR) {
-            newCircle.setFill(new Color(randomGenerator.nextDouble(), randomGenerator.nextDouble(), randomGenerator.nextDouble(), 0.9));
-        } else {
-            newCircle.setFill(new ImagePattern(imageLibrary.pickRandomImage(), 0, 0, 1, 1, true));
-        }
-
-        return newCircle;
-    }
-
-    private void moveCircle(final Circle circle) {
-        final javafx.geometry.Dimension2D dimension2D = gameContext.getGamePanelDimensionProvider().getDimension2D();
-
-        double centerX = 0;
-        double centerY = 0;
-
-        final double timelength = ((maxTimeLength - minTimeLength) * randomGenerator.nextDouble() + minTimeLength) * 1000;
-
-        final Timeline timeline = new Timeline();
-
-        double maxRadius = dimension2D.getHeight()/12;
-
-        if (this.direction == BubblesGameVariant.TOP) {
-            centerX = (dimension2D.getWidth() - maxRadius) * randomGenerator.nextDouble() + maxRadius;
-            centerY = dimension2D.getHeight();
-            timeline.getKeyFrames()
-                .add(new KeyFrame(new Duration(timelength),
-                    new KeyValue(circle.centerYProperty(), -maxRadius, Interpolator.EASE_IN)));
-        } else if (this.direction == BubblesGameVariant.BOTTOM) {
-            centerX = (dimension2D.getWidth() - maxRadius) * randomGenerator.nextDouble() + maxRadius;
-            centerY = 0;
-            timeline.getKeyFrames()
-                .add(new KeyFrame(new Duration(timelength),
-                    new KeyValue(circle.centerYProperty(), dimension2D.getHeight() + maxRadius, Interpolator.EASE_IN)));
-        } else if (this.direction == BubblesGameVariant.RIGHT) {
-            centerX = 0;
-            centerY = (dimension2D.getHeight() - maxRadius) * randomGenerator.nextDouble() + maxRadius;
-            timeline.getKeyFrames()
-                .add(new KeyFrame(new Duration(timelength),
-                    new KeyValue(circle.centerXProperty(), dimension2D.getWidth() + maxRadius, Interpolator.EASE_IN)));
-        } else if (this.direction == BubblesGameVariant.LEFT) {
-            centerX = dimension2D.getWidth();
-            centerY = (dimension2D.getHeight() - maxRadius) * randomGenerator.nextDouble() + maxRadius;
-            timeline.getKeyFrames()
-                .add(new KeyFrame(new Duration(timelength),
-                    new KeyValue(circle.centerXProperty(), -maxRadius, Interpolator.EASE_IN)));
-        }
-
-
-        circle.setCenterX(centerX);
-        circle.setCenterY(centerY);
-
-        timeline.setOnFinished(actionEvent -> {
-            if (this.getChildren().contains(circle)) {
-                gameContext.getGazeDeviceManager().removeEventFilter(circle);
-                this.getChildren().remove(circle);
-                newCircle();
-            }
-        });
-
-        timeline.rateProperty().bind(gameContext.getAnimationSpeedRatioSource().getSpeedRatioProperty());
-
-        timeline.play();
-    }
-
 }
